- Se corrigió el modelo, LineQuantity de PurchaseLine estaba como int, se cambio a decimal. - Se corrigió el metodo Post de CreatePurchase ya que debía actualizar la fecha de los productos comprados sólo si el costo era modificado en la compra.

+++ MinimercadoAlfredo/Models/PurchaseLine.cs
@@ -18,7 +18,7 @@ namespace MinimercadoAlfredo.Models
         [DisplayFormat(DataFormatString = "{0:C2}", ApplyFormatInEditMode = false)]
         public decimal LinePrice { get; set; }
 
-        public int LineQuantity { get; set; }
+        public decimal LineQuantity { get; set; }
 
         [DisplayFormat(DataFormatString = "{0:C2}", ApplyFormatInEditMode = false)]
         public decimal LineTotal { get; set; }
